+++ src/main/java/com/swcamp9th/bangflixbackend/domain/comment/service/CommentServiceImpl.java
@@ -6,8 +6,8 @@ import com.swcamp9th.bangflixbackend.domain.comment.entity.Comment;
 import com.swcamp9th.bangflixbackend.domain.comment.repository.CommentRepository;
 import com.swcamp9th.bangflixbackend.domain.comment.dto.CommentCreateDTO;
 import com.swcamp9th.bangflixbackend.domain.comment.dto.CommentUpdateDTO;
-import com.swcamp9th.bangflixbackend.domain.communityPost.command.entity.CommunityPost;
-import com.swcamp9th.bangflixbackend.domain.communityPost.command.repository.CommunityPostRepository;
+import com.swcamp9th.bangflixbackend.domain.communityPost.entity.CommunityPost;
+import com.swcamp9th.bangflixbackend.domain.communityPost.repository.CommunityPostRepository;
 import com.swcamp9th.bangflixbackend.domain.user.entity.Member;
 import com.swcamp9th.bangflixbackend.domain.user.repository.UserRepository;
 import com.swcamp9th.bangflixbackend.exception.InvalidUserException;
