+++ src/main/java/com/swcamp9th/bangflixbackend/domain/communityPost/dto/CommunityPostCreateDTO.java
@@ -1,4 +1,4 @@
-package com.swcamp9th.bangflixbackend.domain.communityPost.command.dto;
+package com.swcamp9th.bangflixbackend.domain.communityPost.dto;
 
 import lombok.*;
 import org.springframework.web.multipart.MultipartFile;
