+++ src/main/java/com/swcamp9th/bangflixbackend/domain/communityPost/entity/CommunityLikeId.java
@@ -1,4 +1,4 @@
-package com.swcamp9th.bangflixbackend.domain.communityPost.command.entity;
+package com.swcamp9th.bangflixbackend.domain.communityPost.entity;
 
 import lombok.*;
