+++ src/main/java/com/swcamp9th/bangflixbackend/domain/communityPost/entity/CommunityPost.java
@@ -1,4 +1,4 @@
-package com.swcamp9th.bangflixbackend.domain.communityPost.command.entity;
+package com.swcamp9th.bangflixbackend.domain.communityPost.entity;
 
 import com.swcamp9th.bangflixbackend.domain.user.entity.Member;
 import jakarta.persistence.*;
