+++ src/main/java/com/swcamp9th/bangflixbackend/domain/communityPost/repository/CommunityLikeRepository.java
@@ -1,7 +1,7 @@
-package com.swcamp9th.bangflixbackend.domain.communityPost.command.repository;
+package com.swcamp9th.bangflixbackend.domain.communityPost.repository;
 
-import com.swcamp9th.bangflixbackend.domain.communityPost.command.entity.CommunityLike;
-import com.swcamp9th.bangflixbackend.domain.communityPost.command.entity.CommunityLikeId;
+import com.swcamp9th.bangflixbackend.domain.communityPost.entity.CommunityLike;
+import com.swcamp9th.bangflixbackend.domain.communityPost.entity.CommunityLikeId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
