+++ src/main/java/com/swcamp9th/bangflixbackend/domain/communityPost/service/CommunityLikeService.java
@@ -1,0 +1,11 @@
+package com.swcamp9th.bangflixbackend.domain.communityPost.service;
+
+import com.swcamp9th.bangflixbackend.domain.communityPost.dto.CommunityLikeCountDTO;
+import com.swcamp9th.bangflixbackend.domain.communityPost.dto.CommunityLikeCreateDTO;
+
+public interface CommunityLikeService {
+
+    void addLike(String loginId, CommunityLikeCreateDTO newLike);
+
+    CommunityLikeCountDTO countLike(int communityPostCode);
+}
